Renamed Logger.h to Logger.hpp
