+++ Logger/src/SerialLogger.cpp
@@ -8,7 +8,7 @@
 // *****************************************************************************
 // *****************************************************************************
 
-#include "SerialLogger.h"
+#include "SerialLogger.hpp"
 
 SerialLogger::SerialLogger() : Logger()
 {
@@ -19,8 +19,10 @@ SerialLogger::~SerialLogger()
 }
 
 void SerialLogger::log(
+   const LogLevel& logLevel,
    const String& string)
 {
+   Serial.print(toString(logLevel) + " ");
    Serial.print(string);
    Serial.print("\n");
 }
